refactor version logic into the Istio component
